restructure metrics module
